requirements.txt
pinecone-client
python-dotenv
requests
langchain
uvicorn
discord.py
openai
fastapi
pydantic